+++ src/Services/Time/SdxCore.Time.Application/Services/CountryService.cs
@@ -27,7 +27,7 @@ public class CountryService : ICountryService
         return entities.Select(e => SimpleMapper.Map<Country, CountryDto>(e));
     }
 
-    public async Task<CountryDto?> GetByIdAsync(long id, CancellationToken cancellationToken = default) 
+    public async Task<CountryDto?> GetByIdAsync(short id, CancellationToken cancellationToken = default) 
     {
         var entity = await _repository.GetByIdAsync(id, cancellationToken);
         if (entity == null) return null;
@@ -46,7 +46,7 @@ public class CountryService : ICountryService
         return await GetByIdAsync(entity.Id, cancellationToken) ?? throw new InvalidOperationException();
     }
     
-    public async Task<bool> UpdateAsync(long id, UpdateCountryDto dto, CancellationToken cancellationToken = default) 
+    public async Task<bool> UpdateAsync(short id, UpdateCountryDto dto, CancellationToken cancellationToken = default) 
     {
         var entity = await _repository.GetByIdAsync(id, cancellationToken);
         if (entity == null) return false;
@@ -58,7 +58,7 @@ public class CountryService : ICountryService
         return true;
     }
     
-    public async Task<bool> DeleteAsync(long id, CancellationToken cancellationToken = default) 
+    public async Task<bool> DeleteAsync(short id, CancellationToken cancellationToken = default) 
     {
         var entity = await _repository.GetByIdAsync(id, cancellationToken);
         if (entity == null) return false;
